+++ src/js/modules/tab_textures.js
@@ -8,6 +8,7 @@ const ExportHelper = require('../casc/export-helper');
 const EncryptionError = require('../casc/blte-reader').EncryptionError;
 const db2 = require('../casc/db2');
 const textureExporter = require('../ui/texture-exporter');
+const listboxContext = require('../ui/listbox-context');
 const InstallType = require('../install-type');
 
 const texture_atlas_entries = new Map();
@@ -287,7 +288,14 @@ module.exports = {
 		</div>
 		<div class="tab list-tab" id="tab-textures">
 			<div class="list-container">
-				<component :is="$components.Listbox" v-model:selection="$core.view.selectionTextures" :items="$core.view.listfileTextures" :override="$core.view.overrideTextureList" :filter="$core.view.userInputFilterTextures" :keyinput="true" :regex="$core.view.config.regexFilters" :copymode="$core.view.config.copyMode" :pasteselection="$core.view.config.pasteSelection" :copytrimwhitespace="$core.view.config.removePathSpacesCopy" :includefilecount="true" unittype="texture" persistscrollkey="textures"></component>
+				<component :is="$components.Listbox" v-model:selection="$core.view.selectionTextures" :items="$core.view.listfileTextures" :override="$core.view.overrideTextureList" :filter="$core.view.userInputFilterTextures" :keyinput="true" :regex="$core.view.config.regexFilters" :copymode="$core.view.config.copyMode" :pasteselection="$core.view.config.pasteSelection" :copytrimwhitespace="$core.view.config.removePathSpacesCopy" :includefilecount="true" unittype="texture" persistscrollkey="textures" @contextmenu="handle_listbox_context"></component>
+				<component :is="$components.ContextMenu" :node="$core.view.contextMenus.nodeListbox" v-slot:default="context" @close="$core.view.contextMenus.nodeListbox = null">
+					<span @click.self="copy_file_paths(context.node.selection)">Copy file path{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span v-if="context.node.hasFileDataIDs" @click.self="copy_listfile_format(context.node.selection)">Copy file path{{ context.node.count > 1 ? 's' : '' }} (listfile format)</span>
+					<span v-if="context.node.hasFileDataIDs" @click.self="copy_file_data_ids(context.node.selection)">Copy file data ID{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span @click.self="copy_export_paths(context.node.selection)">Copy export path{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span @click.self="open_export_directory(context.node.selection)">Open export directory</span>
+				</component>
 			</div>
 			<div class="filter">
 				<div class="regex-info" v-if="$core.view.config.regexFilters" :title="$core.view.regexTooltip">Regex Enabled</div>
@@ -327,6 +335,30 @@ module.exports = {
 	},
 
 	methods: {
+		handle_listbox_context(data) {
+			listboxContext.handle_context_menu(data);
+		},
+
+		copy_file_paths(selection) {
+			listboxContext.copy_file_paths(selection);
+		},
+
+		copy_listfile_format(selection) {
+			listboxContext.copy_listfile_format(selection);
+		},
+
+		copy_file_data_ids(selection) {
+			listboxContext.copy_file_data_ids(selection);
+		},
+
+		copy_export_paths(selection) {
+			listboxContext.copy_export_paths(selection);
+		},
+
+		open_export_directory(selection) {
+			listboxContext.open_export_directory(selection);
+		},
+
 		is_baked_npc_texture() {
 			return is_baked_npc_texture(this.$core);
 		},