+++ src/js/modules/tab_videos.js
@@ -11,6 +11,7 @@ const InstallType = require('../install-type');
 const constants = require('../constants');
 const core = require('../core');
 const subtitles = require('../subtitles');
+const listboxContext = require('../ui/listbox-context');
 const { BlobPolyfill, URLPolyfill } = require('../blob');
 
 let movie_variation_map = null;
@@ -482,7 +483,14 @@ module.exports = {
 	template: `
 		<div class="tab list-tab" id="tab-video">
 			<div class="list-container">
-				<component :is="$components.Listbox" v-model:selection="$core.view.selectionVideos" :items="$core.view.listfileVideos" :filter="$core.view.userInputFilterVideos" :keyinput="true" :regex="$core.view.config.regexFilters" :copymode="$core.view.config.copyMode" :pasteselection="$core.view.config.pasteSelection" :copytrimwhitespace="$core.view.config.removePathSpacesCopy" :includefilecount="true" unittype="video" persistscrollkey="videos"></component>
+				<component :is="$components.Listbox" v-model:selection="$core.view.selectionVideos" :items="$core.view.listfileVideos" :filter="$core.view.userInputFilterVideos" :keyinput="true" :regex="$core.view.config.regexFilters" :copymode="$core.view.config.copyMode" :pasteselection="$core.view.config.pasteSelection" :copytrimwhitespace="$core.view.config.removePathSpacesCopy" :includefilecount="true" unittype="video" persistscrollkey="videos" @contextmenu="handle_listbox_context"></component>
+				<component :is="$components.ContextMenu" :node="$core.view.contextMenus.nodeListbox" v-slot:default="context" @close="$core.view.contextMenus.nodeListbox = null">
+					<span @click.self="copy_file_paths(context.node.selection)">Copy file path{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span v-if="context.node.hasFileDataIDs" @click.self="copy_listfile_format(context.node.selection)">Copy file path{{ context.node.count > 1 ? 's' : '' }} (listfile format)</span>
+					<span v-if="context.node.hasFileDataIDs" @click.self="copy_file_data_ids(context.node.selection)">Copy file data ID{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span @click.self="copy_export_paths(context.node.selection)">Copy export path{{ context.node.count > 1 ? 's' : '' }}</span>
+					<span @click.self="open_export_directory(context.node.selection)">Open export directory</span>
+				</component>
 			</div>
 			<div class="filter">
 				<div class="regex-info" v-if="$core.view.config.regexFilters" :title="$core.view.regexTooltip">Regex Enabled</div>
@@ -507,6 +515,30 @@ module.exports = {
 	`,
 
 	methods: {
+		handle_listbox_context(data) {
+			listboxContext.handle_context_menu(data);
+		},
+
+		copy_file_paths(selection) {
+			listboxContext.copy_file_paths(selection);
+		},
+
+		copy_listfile_format(selection) {
+			listboxContext.copy_listfile_format(selection);
+		},
+
+		copy_file_data_ids(selection) {
+			listboxContext.copy_file_data_ids(selection);
+		},
+
+		copy_export_paths(selection) {
+			listboxContext.copy_export_paths(selection);
+		},
+
+		open_export_directory(selection) {
+			listboxContext.open_export_directory(selection);
+		},
+
 		async preview_video() {
 			if (this.$core.view.videoPlayerState) {
 				await stop_video(this.$core);